pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "page_analyzer (project 3 on hexlet)"
authors = ["Work Chachimin <chachimin@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "^2.9.9"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
gunicorn = "^23.0.0"
flask = "^3.0.3"
python-dotenv = "^1.0.1"
validators = "^0.34.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
